(refactoring) Converts HTML generator to recursive style
This allows users to subclass the HTML generator and
override methods in a per-production basis, instead
of having to deal with the private @html attribute.

The root Walker was also changed to use the Array#map
method instead of Array#each, that does pretty much the
same thing, but keeps the results and allows posterior
processing.


git-svn-id: http://mediacloth.rubyforge.org/svn/trunk@39 ed6bb6d2-8713-0410-a66e-f7c184f9b1ef