Adding basic support for tables. Unlike with some other constructs, the lexer isn't as good as figuring out if needs to close an unopen table. This will require some major surgery at a later date.
The current implementation treats the table options as a simple string.
This should probably be enhanced, with the lexer splitting the string
into name/value pairs. However, I  have left this for later, as we will
probably need to do this to support general (X)HTML markup.

The table grammar is pretty flexible: head cells and body cells are
allowed to be mixed in a row, as Mediawiki allows this, and browsers
seem to allow this, as well.

Also adding a new test file to test a variety of table examples.



git-svn-id: http://mediacloth.rubyforge.org/svn/trunk@67 ed6bb6d2-8713-0410-a66e-f7c184f9b1ef